pyproject.toml
[tool.poetry]
name = "hodolbot"
version = "0.1.0"
description = ""
authors = ["solar0037 <solar0037@gmail.com>"]

[tool.poetry.scripts]
start = "hodolbot.app:main"

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.0"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
